RF: Updated the window file
- Flatten the window folder to the file.
- Added all the pygfx imports from window file to lib.
- Created jupyter-rfb as optional dependency under plot.
