--- fury/window/__init__.py
@@ -1,3 +1,0 @@
-import lazy_loader as lazy
-
-__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)
