--- fury/window/__init__.pyi
@@ -1,21 +1,0 @@
-__all__ = [
-    "ShowManager",
-    "Scene",
-    "record",
-    "update_camera",
-    "update_viewports",
-    "render_screens",
-    "calculate_screen_sizes",
-    "create_screen",
-    "display",
-]
-
-from .screen import (
-    Scene,
-    calculate_screen_sizes,
-    create_screen,
-    render_screens,
-    update_camera,
-    update_viewports,
-)
-from .show_manager import ShowManager, display, record
